server/requirements.txt
Django==4.1.3
Pillow==9.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
APScheduler==3.9.1.post1
django-cors-headers==3.13.0
